remove utils dependencies in whitelist
